services/maia/pyproject.toml
[project]
name = "chessbeast-maia"
version = "0.1.0"
description = "Maia gRPC service for ChessBeast"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.60",
    "protobuf>=4.25",
    "chess>=1.10",
    "maia2>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/maia_service"]
